+++ game/scripts/vscripts/util.ts
@@ -69,3 +69,41 @@ export function setUnitPacifist(unit: CDOTA_BaseNPC, isPacifist: boolean, durati
         unit.RemoveModifierByName("modifier_tutorial_pacifist");
     }
 }
+
+/**
+ * Moves the camera to a unit's position. The player can't control the camera
+ * while it is being moved.
+ * @param target Unit to move the camera to.
+ * @param duration Time in seconds it takes to move the camera
+ */
+export const moveCameraToUnit = (target: CBaseEntity, duration: number) => {
+    let playerId = findRealPlayerID();
+    let player = PlayerResource.GetPlayer(playerId);
+  
+    if (player) {
+        CustomGameEventManager.Send_ServerToPlayer(player, "move_camera", {
+            unitTargetEntIndex: target.GetEntityIndex(),
+            lerp: duration
+        });
+    }
+  }
+  
+  /**
+  * Moves the camera to a position. The player can't control the camera
+  * while it is being moved.
+  * @param position Point to move the camera to.
+  * @param duration Time in seconds it takes to move the camera
+  */
+  export const moveCameraToPosition = (position: Vector, duration: number) => {
+    let playerId = findRealPlayerID();
+    let player = PlayerResource.GetPlayer(playerId);
+    
+    if (player) {
+        CustomGameEventManager.Send_ServerToPlayer(player, "move_camera", {
+            cameraTargetX: position.x,
+            cameraTargetY: position.y,
+            cameraTargetZ: position.z,
+            lerp: duration
+        });
+    }
+  }
